move unit test to destination_test.go
